refactor: rename configuration to configurationSingleton
