+++ src/ConfigurationSingleton.ts
